refactor: re-分割
- Added BackgroundIntent.swift to handle background tasks for the widget.
- Implemented URL parameter handling and logging for better debugging.
- Enhanced widget update logic to ensure tasks are processed correctly in the background.
